createDbiConnectionDetails {DatabaseConnector}	R Documentation

Create DBI connection details

Description

Create DBI connection details

Usage

createDbiConnectionDetails(dbms, drv, ...)

Arguments

dbms
The type of DBMS running on the server. Valid values are

"oracle" for Oracle

"postgresql" for PostgreSQL

"redshift" for Amazon Redshift

"sql server" for Microsoft SQL Server

"pdw" for Microsoft Parallel Data Warehouse (PDW)

"netezza" for IBM Netezza

"bigquery" for Google BigQuery

"sqlite" for SQLite

"sqlite extended" for SQLite with extended types (DATE and DATETIME)

"spark" for Spark

"snowflake" for Snowflake

drv
An object that inherits from DBIDriver, or an existing DBIConnection object (in order to clone an existing connection).

...
authentication arguments needed by the DBMS instance; these typically include user, password, host, port, dbname, etc. For details see the appropriate DBIDriver

Value

A list with all the details needed to connect to a database.

[Package DatabaseConnector version 6.0.0 Index]