dateAdd {DatabaseConnector}	R Documentation

Add an interval to a date

Description

This function is provided primarily to be used together with dbplyr when querying a database. It will also work in dplyr against data frames.

Usage

dateAdd(interval, number, date)

Arguments

interval
Unit for the interval. Can be "day", "week", "month", "year".

number
The number of units to add to the date.

date
The date to add to.

Value

A new date.

Examples

dateAdd("day", 10, as.Date("2000-01-01"))

[Package DatabaseConnector version 6.3.2 Index]