dateFromParts {DatabaseConnector}	R Documentation

Construct a date from parts

Description

This function is provided primarily to be used together with dbplyr when querying a database. It will also work in dplyr against data frames.

Usage

dateFromParts(year, month, day)

Arguments

year
The calendar year.

month
The calendar month (1 = January).

day
The day of the month.

Value

The date.

Examples

dateFromParts(2000, 1, 5)

[Package DatabaseConnector version 6.3.2 Index]