Insert rows into a table

The dbAppendTable() method assumes that the table has been created beforehand, e.g. with dbCreateTable(). The default implementation calls sqlAppendTableTemplate() and then dbExecute() with the param argument. Backends compliant to ANSI SQL 99 which use ? as a placeholder for prepard queries don't need to override it. Backends with a different SQL syntax which use ? as a placeholder for prepared queries can override sqlAppendTable(). Other backends (with different placeholders or with entirely different ways to create tables) need to override the dbAppendTable() method.

# S4 method for DatabaseConnectorConnection,character,data.frame
dbAppendTable(conn,
  name, value, temporary = FALSE, oracleTempSchema = NULL, ...,
  row.names = NULL)

Arguments

conn
A DBIConnection object, as returned by dbConnect().

name
Name of the table, escaped with dbQuoteIdentifier().

value
A data frame of values. The column names must be consistent with those in the target table in the database.

temporary
Should the table created as a temp table?

oracleTempSchema
Specifically for Oracle, a schema with write priviliges where temp tables can be created.

...
Other arguments used by individual methods.

row.names
Must be NULL.

Details

The row.names argument is not supported by this method. Process the values with sqlRownamesToColumn() before calling this method.

See also