dbClearResult,DatabaseConnectorDbiResult-method {DatabaseConnector}	R Documentation

Clear a result set

Description

Frees all resources (local and remote) associated with a result set. This step is mandatory for all objects obtained by calling dbSendQuery() or dbSendStatement().

Usage

## S4 method for signature 'DatabaseConnectorDbiResult'
dbClearResult(res, ...)

Arguments

res
An object inheriting from DBIResult.

...
Other arguments passed on to methods.

Value

dbClearResult() returns TRUE, invisibly, for result sets obtained from both dbSendQuery() and dbSendStatement().

See Also

Other DBIResult generics: DBIResult-class, dbBind(), dbColumnInfo(), dbFetch(), dbGetInfo(), dbGetRowCount(), dbGetRowsAffected(), dbGetStatement(), dbHasCompleted(), dbIsReadOnly(), dbIsValid(), dbQuoteIdentifier(), dbQuoteLiteral(), dbQuoteString(), dbUnquoteIdentifier()

[Package DatabaseConnector version 6.3.2 Index]