dbConnect,DatabaseConnectorDriver-method {DatabaseConnector}	R Documentation

Create a connection to a DBMS

Description

Connect to a database. This function is synonymous with the connect() function. except a dummy driver needs to be specified

Usage

## S4 method for signature 'DatabaseConnectorDriver'
dbConnect(drv, ...)

Arguments

drv
The result of the DatabaseConnectorDriver() function

...
Other parameters. These are the same as expected by the connect() function.

Value

Returns a DatabaseConnectorConnection object that can be used with most of the other functions in this package.

Examples

## Not run: 
conn <- dbConnect(DatabaseConnectorDriver(),
  dbms = "postgresql",
  server = "localhost/ohdsi",
  user = "joe",
  password = "secret"
)
querySql(conn, "SELECT * FROM cdm_synpuf.person;")
dbDisconnect(conn)

## End(Not run)

[Package DatabaseConnector version 6.3.2 Index]