Fetch records from a previously executed query

Fetch the next n elements (rows) from the result set and return them as a data.frame.

# S4 method for DatabaseConnectorJdbcResult
dbFetch(res, n = -1, ...)

Arguments

res

An object inheriting from DBIResult, created by dbSendQuery().

n

maximum number of records to retrieve per fetch. Use n = -1 or n = Inf to retrieve all pending records. Some implementations may recognize other special values.

...

Other arguments passed on to methods.

Value

dbFetch() always returns a data.frame

with as many rows as records were fetched and as many columns as fields in the result set, even if the result is a single value or has one or zero rows.

Details

fetch() is provided for compatibility with older DBI clients - for all new code you are strongly encouraged to use dbFetch(). The default implementation for dbFetch() calls fetch() so that it is compatible with existing code. Modern backends should implement for dbFetch() only.

See also