List field names of a remote table

List field names of a remote table

# S4 method for DatabaseConnectorConnection,character
dbListFields(conn, name, database = NULL, schema = NULL, ...)

Arguments

conn
A DBIConnection object, as returned by dbConnect().

name
a character string with the name of the remote table.

database
Name of the database.

schema
Name of the schema.

...
Other parameters passed on to methods.

Value

dbListFields() returns a character vector that enumerates all fields in the table in the correct order. This also works for temporary tables if supported by the database. The returned names are suitable for quoting with dbQuoteIdentifier(). If the table does not exist, an error is raised. Invalid types for the name argument (e.g., character of length not equal to one, or numeric) lead to an error. An error is also raised when calling this method for a closed or invalid connection.

See also