dbRemoveTable,DatabaseConnectorConnection,character-method {DatabaseConnector}	R Documentation

Remove a table from the database

Description

Remove a remote table (e.g., created by dbWriteTable()) from the database.

Usage

## S4 method for signature 'DatabaseConnectorConnection,character'
dbRemoveTable(
  conn,
  name,
  database = NULL,
  schema = NULL,
  oracleTempSchema = NULL,
  ...
)

Arguments

conn
A DBIConnection object, as returned by dbConnect().

name
A character string specifying a DBMS table name.

database
Name of the database.

schema
Name of the schema.

oracleTempSchema
Specifically for Oracle, a schema with write privileges where temp tables can be created.

...
Other parameters passed on to methods.

Value

dbRemoveTable() returns TRUE, invisibly. If the table does not exist, an error is raised. An attempt to remove a view with this function may result in an error.

An error is raised when calling this method for a closed or invalid connection. An error is also raised if name cannot be processed with dbQuoteIdentifier() or if this results in a non-scalar.

See Also

Other DBIConnection generics: DBIConnection-class, dbAppendTable(), dbCreateTable(), dbDataType(), dbDisconnect(), dbExecute(), dbExistsTable(), dbGetException(), dbGetInfo(), dbGetQuery(), dbIsReadOnly(), dbIsValid(), dbListFields(), dbListObjects(), dbListResults(), dbListTables(), dbReadTable(), dbSendQuery(), dbSendStatement(), dbWriteTable()

[Package DatabaseConnector version 4.0.2 Index]