Download DatabaseConnector JDBC Jar files

Download the DatabaseConnector JDBC drivers from https://ohdsi.github.io/DatabaseConnectorJars/

downloadJdbcDrivers(
  dbms,
  pathToDriver = Sys.getenv("DATABASECONNECTOR_JAR_FOLDER"),
  method = "auto",
  ...
)

Arguments

dbms
The type of DBMS to download Jar files for.

"postgresql" for PostgreSQL

"redshift" for Amazon Redshift

"sql server" or "pdw" for Microsoft SQL Server

"oracle" for Oracle

"spark" for Spark

pathToDriver
The full path to the folder where the JDBC driver .jar files should be downloaded to. By default the value of the environment variable "DATABASECONNECTOR_JAR_FOLDER" is used.

method
The method used for downloading files. See ?download.file for details and options.

...
Further arguments passed on to download.file

Value

Invisibly returns the destination if the download was successful.

Details

The following versions of the JDBC drivers are currently used:

PostgreSQLV42.2.18

RedShiftV1.2.27.1051

SQL ServerV8.4.1.zip

OracleV19.8

SparkV2.6.17

Examples

if (FALSE) {
downloadJdbcDrivers("redshift")
}