getTableNames {DatabaseConnector}	R Documentation

List all tables in a database schema.

Description

This function returns a list of all tables in a database schema.

Usage

getTableNames(connection, databaseSchema)

Arguments

connection
The connection to the database server.

databaseSchema
The name of the database schema. See details for platform-specific details.

Details

The databaseSchema argument is interpreted differently according to the different platforms: SQL Server and PDW: The databaseSchema schema should specify both the database and the schema, e.g. 'my_database.dbo'. PostgreSQL and Redshift: The databaseSchema should specify the schema. Oracle: The databaseSchema should specify the Oracle 'user'. MySql and Impala: The databaseSchema should specify the database.

Value

A character vector of table names. To ensure consistency across platforms, these table names are in upper case.

[Package DatabaseConnector version 4.0.2 Index]