inDatabaseSchema {DatabaseConnector}	R Documentation

Refer to a table in a database schema

Description

Can be used with dplyr::tbl() to indicate a table in a specific database schema.

Usage

inDatabaseSchema(databaseSchema, table)

Arguments

databaseSchema
The name of the database schema. See details for platform-specific details.

table
The name of the table in the database schema.

Details

The databaseSchema argument is interpreted differently according to the different platforms: SQL Server and PDW: The databaseSchema schema should specify both the database and the schema, e.g. 'my_database.dbo'. Impala: the databaseSchema should specify the database. Oracle: The databaseSchema should specify the Oracle 'user'. All other : The databaseSchema should specify the schema.

Value

An object representing the table and database schema.

[Package DatabaseConnector version 6.3.2 Index]