isSqlReservedWord {DatabaseConnector}	R Documentation

Test a character vector of SQL names for SQL reserved words

Description

This function checks a character vector against a predefined list of reserved SQL words.

Usage

isSqlReservedWord(sqlNames, warn = FALSE)

Arguments

sqlNames
A character vector containing table or field names to check.

warn
(logical) Should a warn be thrown if invalid SQL names are found?

Value

A logical vector with length equal to sqlNames that is TRUE for each name that is reserved and FALSE otherwise

[Package DatabaseConnector version 6.3.2 Index]