lowLevelQuerySql {DatabaseConnector}	R Documentation

Low level function for retrieving data to a data frame

Description

This is the equivalent of the querySql() function, except no error report is written when an error occurs.

Usage

lowLevelQuerySql(
  connection,
  query,
  datesAsString = FALSE,
  integerAsNumeric = getOption("databaseConnectorIntegerAsNumeric", default = TRUE),
  integer64AsNumeric = getOption("databaseConnectorInteger64AsNumeric", default = TRUE)
)

Arguments

connection
The connection to the database server created using either connect() or dbConnect().

query
The SQL statement to retrieve the data

datesAsString
Logical: Should dates be imported as character vectors, our should they be converted to R's date format?

integerAsNumeric
Logical: should 32-bit integers be converted to numeric (double) values? If FALSE 32-bit integers will be represented using R's native Integer class.

integer64AsNumeric
Logical: should 64-bit integers be converted to numeric (double) values? If FALSE 64-bit integers will be represented using bit64::integer64.

Details

Retrieves data from the database server and stores it in a data frame. Null values in the database are converted to NA values in R.

Value

A data frame containing the data retrieved from the server

[Package DatabaseConnector version 6.3.2 Index]