lowLevelQuerySqlToAndromeda {DatabaseConnector}	R Documentation

Low level function for retrieving data to a local Andromeda object

Description

This is the equivalent of the querySqlToAndromeda() function, except no error report is written when an error occurs.

Usage

lowLevelQuerySqlToAndromeda(
  connection,
  query,
  andromeda,
  andromedaTableName,
  datesAsString = FALSE,
  appendToTable = FALSE,
  snakeCaseToCamelCase = FALSE,
  integerAsNumeric = getOption("databaseConnectorIntegerAsNumeric", default = TRUE),
  integer64AsNumeric = getOption("databaseConnectorInteger64AsNumeric", default = TRUE)
)

Arguments

connection
The connection to the database server created using either connect() or dbConnect().

query
The SQL statement to retrieve the data

andromeda
An open Andromeda object, for example as created using Andromeda::andromeda().

andromedaTableName
The name of the table in the local Andromeda object where the results of the query will be stored.

datesAsString
Should dates be imported as character vectors, our should they be converted to R's date format?

appendToTable
If FALSE, any existing table in the Andromeda with the same name will be replaced with the new data. If TRUE, data will be appended to an existing table, assuming it has the exact same structure.

snakeCaseToCamelCase
If true, field names are assumed to use snake_case, and are converted to camelCase.

integerAsNumeric
Logical: should 32-bit integers be converted to numeric (double) values? If FALSE 32-bit integers will be represented using R's native Integer class.

integer64AsNumeric
Logical: should 64-bit integers be converted to numeric (double) values? If FALSE 64-bit integers will be represented using bit64::integer64.

Details

Retrieves data from the database server and stores it in a local Andromeda object This allows very large data sets to be retrieved without running out of memory. Null values in the database are converted to NA values in R. If a table with the same name already exists in the local Andromeda object it is replaced.

Value

Invisibly returns the andromeda. The Andromeda object will have a table added with the query results.

[Package DatabaseConnector version 6.3.2 Index]