year {DatabaseConnector}	R Documentation

Extract the year from a date

Description

This function is provided primarily to be used together with dbplyr when querying a database. It will also work in dplyr against data frames.

Usage

year(date)

Arguments

date
The date.

Value

The year

Examples

year(as.Date("2000-02-01"))

[Package DatabaseConnector version 6.3.2 Index]