DatabaseConnectorConnection-class {DatabaseConnector}	R Documentation

DatabaseConnectorConnection class.

Description

DatabaseConnectorConnection class.

Usage

## S4 method for signature 'DatabaseConnectorConnection'
dbDisconnect(conn)

## S4 method for signature 'DatabaseConnectorConnection'
show(object)

## S4 method for signature 'DatabaseConnectorConnection,character'
dbQuoteIdentifier(conn, x, ...)

## S4 method for signature 'DatabaseConnectorConnection,character'
dbQuoteString(conn, x, ...)

## S4 method for signature 'DatabaseConnectorConnection,character'
dbGetQuery(conn, statement, ...)

## S4 method for signature 'DatabaseConnectorConnection,character'
dbSendStatement(conn, statement, ...)

## S4 method for signature 'DatabaseConnectorConnection,character'
dbExecute(conn, statement, ...)

## S4 method for signature 'DatabaseConnectorConnection,character'
dbListFields(conn, name, database = NULL, schema = NULL, ...)

## S4 method for signature 'DatabaseConnectorConnection'
dbListTables(conn, database = NULL, schema = NULL, ...)

## S4 method for signature 'DatabaseConnectorConnection,character'
dbExistsTable(conn, name, database = NULL, schema = NULL, ...)

## S4 method for signature 'DatabaseConnectorConnection,ANY'
dbWriteTable(
  conn,
  name,
  value,
  overwrite = FALSE,
  append = FALSE,
  temporary = FALSE,
  oracleTempSchema = NULL,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  ...
)

## S4 method for signature 'DatabaseConnectorConnection,character'
dbAppendTable(
  conn,
  name,
  value,
  temporary = FALSE,
  oracleTempSchema = NULL,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  ...,
  row.names = NULL
)

## S4 method for signature 'DatabaseConnectorConnection'
dbCreateTable(
  conn,
  name,
  fields,
  oracleTempSchema = NULL,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  ...,
  row.names = NULL,
  temporary = FALSE
)

## S4 method for signature 'DatabaseConnectorConnection,character'
dbReadTable(
  conn,
  name,
  database = NULL,
  schema = NULL,
  oracleTempSchema = NULL,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  ...
)

## S4 method for signature 'DatabaseConnectorConnection,ANY'
dbRemoveTable(
  conn,
  name,
  database = NULL,
  schema = NULL,
  oracleTempSchema = NULL,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  ...
)

Arguments

database
Name of the database.

schema
Name of the schema.

overwrite
Overwrite an existing table (if exists)?

append
Append to existing table?

temporary
Should the table created as a temp table?

oracleTempSchema
DEPRECATED: use tempEmulationSchema instead.

tempEmulationSchema
Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp tables, provide a schema with write privileges where temp tables can be created.

[Package DatabaseConnector version 5.1.0 Index]