assertTempEmulationSchemaSet {DatabaseConnector}	R Documentation

Assert the temp emulation schema is set

Description

Asserts the temp emulation schema is set for DBMSs requiring temp table emulation.

If you know your code uses temp tables, it is a good idea to call this function first, so it can throw an informative error if the user forgot to set the temp emulation schema.

Usage

assertTempEmulationSchemaSet(
  dbms,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema")
)

Arguments

dbms
The type of DBMS running on the server. See connect() or createConnectionDetails() for valid values.

tempEmulationSchema
The temp emulation schema specified by the user.

Value

Does not return anything. Throws an error if the DBMS requires temp emulation but the temp emulation schema is not set.

[Package DatabaseConnector version 6.3.2 Index]