computeDataHash {DatabaseConnector}	R Documentation

Compute hash of data

Description

Compute a hash of the data in the database schema. If the data changes, this should produce a different hash code. Specifically, the hash is based on the field names, field types, and table row counts.

Usage

computeDataHash(connection, databaseSchema, tables = NULL, progressBar = TRUE)

Arguments

connection
The connection to the database server created using either connect() or dbConnect().

databaseSchema
The name of the database schema. See details for platform-specific details.

tables
(Optional) A list of tables to restrict to.

progressBar
When true, a progress bar is shown based on the number of tables in the database schema.

Details

The databaseSchema argument is interpreted differently according to the different platforms: SQL Server and PDW: The databaseSchema schema should specify both the database and the schema, e.g. 'my_database.dbo'. Impala: the databaseSchema should specify the database. Oracle: The databaseSchema should specify the Oracle 'user'. All other : The databaseSchema should specify the schema.

Value

A string representing the MD5 hash code.

[Package DatabaseConnector version 6.3.2 Index]