createZipFile {DatabaseConnector}	R Documentation

Compress files and/or folders into a single zip file

Description

Compress files and/or folders into a single zip file

Usage

createZipFile(zipFile, files, rootFolder = getwd(), compressionLevel = 9)

Arguments

zipFile
The path to the zip file to be created.

files
The files and/or folders to be included in the zip file. Folders will be included recursively.

rootFolder
The root folder. All files will be stored with relative paths relative to this folder.

compressionLevel
A number between 1 and 9. 9 compresses best, but it also takes the longest.

Details

Uses Java's compression library to create a zip file. It is similar to utils::zip, except that it does not require an external zip tool to be available on the system path.

[Package DatabaseConnector version 6.3.2 Index]