dbCreateTable,DatabaseConnectorConnection,character,data.frame-method {DatabaseConnector}	R Documentation

Create a table in the database

Description

The default dbCreateTable() method calls sqlCreateTable() and dbExecute(). Backends compliant to ANSI SQL 99 don't need to override it. Backends with a different SQL syntax can override sqlCreateTable(), backends with entirely different ways to create tables need to override this method.

Usage

## S4 method for signature 'DatabaseConnectorConnection,character,data.frame'
dbCreateTable(
  conn,
  name,
  fields,
  oracleTempSchema = NULL,
  ...,
  row.names = NULL,
  temporary = FALSE
)

Arguments

conn
A DBIConnection object, as returned by dbConnect().

name
Name of the table, escaped with dbQuoteIdentifier().

fields
Either a character vector or a data frame.

A named character vector: Names are column names, values are types. Names are escaped with dbQuoteIdentifier(). Field types are unescaped.

A data frame: field types are generated using dbDataType().

oracleTempSchema
Specifically for Oracle, a schema with write privileges where temp tables can be created.

...
Other arguments used by individual methods.

row.names
Must be NULL.

temporary
Should the table created as a temp table?

Details

The row.names argument is not supported by this method. Process the values with sqlRownamesToColumn() before calling this method.

The argument order is different from the sqlCreateTable() method, the latter will be adapted in a later release of DBI.

See Also

Other DBIConnection generics: DBIConnection-class, dbAppendTable(), dbDataType(), dbDisconnect(), dbExecute(), dbExistsTable(), dbGetException(), dbGetInfo(), dbGetQuery(), dbIsReadOnly(), dbIsValid(), dbListFields(), dbListObjects(), dbListResults(), dbListTables(), dbReadTable(), dbRemoveTable(), dbSendQuery(), dbSendStatement(), dbWriteTable()

[Package DatabaseConnector version 4.0.2 Index]