dbms {DatabaseConnector}	R Documentation

Get the database platform from a connection

Description

The SqlRender package provides functions that translate SQL from OHDSI-SQL to a target SQL dialect. These function need the name of the database platform to translate to. The dbms function returns the dbms for any DBI connection that can be passed along to SqlRender translation functions (see example).

Usage

dbms(connection)

Arguments

connection
The connection to the database server created using either connect() or dbConnect().

Value

The name of the database (dbms) used by SqlRender

Examples

library(DatabaseConnector)
con <- connect(dbms = "sqlite", server = ":memory:")
dbms(con)
#> [1] "sqlite"
SqlRender::translate("DATEADD(d, 365, dateColumn)", targetDialect = dbms(con))
#> "CAST(STRFTIME('%s', DATETIME(dateColumn, 'unixepoch', (365)||' days')) AS REAL)"
disconnect(con)

[Package DatabaseConnector version 6.3.2 Index]