downloadJdbcDrivers {DatabaseConnector}	R Documentation

Download DatabaseConnector JDBC Jar files

Description

Download the DatabaseConnector JDBC drivers from https://ohdsi.github.io/DatabaseConnectorJars/

Usage

downloadJdbcDrivers(
  dbms,
  pathToDriver = Sys.getenv("DATABASECONNECTOR_JAR_FOLDER"),
  method = "auto",
  ...
)

Arguments

dbms
The type of DBMS to download Jar files for.

"postgresql" for PostgreSQL

"redshift" for Amazon Redshift

"sql server", "pdw" or "synapse" for Microsoft SQL Server

"oracle" for Oracle

"spark" for Spark

"snowflake" for Snowflake

"bigquery" for Google BigQuery

"all" for all aforementioned platforms

pathToDriver
The full path to the folder where the JDBC driver .jar files should be downloaded to. By default the value of the environment variable "DATABASECONNECTOR_JAR_FOLDER" is used.

method
The method used for downloading files. See ?download.file for details and options.

...
Further arguments passed on to download.file.

Details

The following versions of the JDBC drivers are currently used:

PostgreSQL: V42.2.18

RedShift: V2.1.0.9

SQL Server: V9.2.0

Oracle: V19.8

Spark: V2.6.21

Snowflake: V3.13.22

BigQuery: v1.3.2.1003

Value

Invisibly returns the destination if the download was successful.

Examples

## Not run: 
downloadJdbcDrivers("redshift")

## End(Not run)

[Package DatabaseConnector version 6.3.2 Index]