How to download and use JDBC drivers for the various data platforms.

Below are instructions for downloading JDBC drivers for the various data platforms. Once downloaded use the pathToDriver argument in the connect or createConnectionDetails functions to point to the driver.

PostgresSql

Go to the PostgresSQL JDBC site and download the current version. The file is called something like 'postgresql-42.2.2.jar'.

Oracle

Go to the Oracle JDBC site. Select 'Accept License Agreement' and download the jar file. The file is called something like 'ojdbc7.jar'.

SQL Server and PDW

Go to the Microsoft SQL Server JDBC site, click 'Download' and select the tar.gz file. Click 'Next' to start the download. Decompress the file and find a file called something like 'sqljdbc41.jar' in the a folder named something like 'sqljdbc_6.0/enu/jre7'.

RedShift

Go to the Amazon RedShfit JDBC driver page and download the latest JDBC driver. The file is called something like 'RedshiftJDBC42-1.2.12.1017.jar'.

Netezza

Read the instructions here on how to obtain the Netezza JDBC driver.

BigQuery

Go to Google's site and download the latest JDBC driver. Unzip the file, and locate the appropriate jar files.

Impala

Go to Cloudera's site, pick your OS version, and click "GET IT NOW!'. Register, and you should be able to download the driver.

Spark

Go to Simba Spark site and download the latest JDBC driver (license may be required).